Segunda correcion sexto punto

+++ 6.cpp
@@ -5,7 +5,7 @@ int main (void)
 {
   std::cout.precision(16);
   std::cout.setf(std::ios::scientific);
-  long double sum=0;
+  double sum=0;
   for(double ii=1; ii<1000000; ii++)
     {
       sum+=1/ii;
